+++ CHANGELOG.md
@@ -10,6 +10,9 @@
   [#37](https://github.com/kylef/Commander/issues/37)
 - Argument descriptions are now printed in command help.
   [#33](https://github.com/kylef/Commander/issues/33)
+- Default option and flag default values will now be shown in help output.
+  Only default option types of String and Int are currently supported in help output.
+  [#34](https://github.com/kylef/Commander/issues/34)
 
 ### Bug Fixes